requirements.txt
discord.py>=2.3.0
yt-dlp>=2023.3.4
PyNaCl>=1.5.0
aiohttp>=3.9.0